added NoMoveService, UniqueService and SharedService
