+++ CMakeLists.txt
@@ -1,5 +1,4 @@
 cmake_minimum_required(VERSION 2.6)
 project(kangaru)
 
-set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/bin)
 add_subdirectory(examples)
